Move matrix and vector to separate files
