+++ src/asgard_dimension.hpp
@@ -13,9 +13,10 @@
 
 #include "fast_math.hpp"
 #include "matlab_utilities.hpp"
+#include "matrix.hpp"
 #include "program_options.hpp"
-#include "tensors.hpp"
 #include "tools.hpp"
+#include "vector.hpp"
 
 namespace asgard
 {
