+++ src/asgard_kronmult_matrix.hpp
@@ -7,6 +7,7 @@
 #include "distribution.hpp"
 #include "elements.hpp"
 #include "pde.hpp"
+#include "vector.hpp"
 
 #include "./device/asgard_kronmult.hpp"
