+++ src/basis.cpp
@@ -2,9 +2,10 @@
 #include "distribution.hpp"
 #include "fast_math.hpp"
 #include "matlab_utilities.hpp"
+#include "matrix.hpp"
 #include "quadrature.hpp"
-#include "tensors.hpp"
 #include "tools.hpp"
+#include "vector.hpp"
 #include <algorithm>
 #include <cmath>
 #include <numeric>
