+++ src/batch.hpp
@@ -1,8 +1,9 @@
 #pragma once
 #include "elements.hpp"
+#include "matrix.hpp"
 #include "pde/pde_base.hpp"
-#include "tensors.hpp"
 #include "tools.hpp"
+#include "vector.hpp"
 #include <array>
 #include <numeric>
