+++ src/coefficients_tests.cpp
@@ -1,10 +1,11 @@
 #include "basis.hpp"
 #include "coefficients.hpp"
 #include "matlab_utilities.hpp"
+#include "matrix.hpp"
 #include "pde.hpp"
 #include "pde/pde_base.hpp"
-#include "tensors.hpp"
 #include "tests_general.hpp"
+#include "vector.hpp"
 
 static auto const coefficients_base_dir = gold_base_dir / "coefficients";
