+++ src/elements.cpp
@@ -2,9 +2,10 @@
 
 #include "fast_math.hpp"
 #include "matlab_utilities.hpp"
+#include "matrix.hpp"
 #include "program_options.hpp"
-#include "tensors.hpp"
 #include "tools.hpp"
+#include "vector.hpp"
 
 #include <cmath>
 #include <functional>
