+++ src/fast_math.hpp
@@ -1,9 +1,10 @@
 #pragma once
 #include "lib_dispatch.hpp"
+#include "matrix.hpp"
 #include "program_options.hpp"
 #include "sparse.hpp"
-#include "tensors.hpp"
 #include "tools.hpp"
+#include "vector.hpp"
 #ifdef ASGARD_USE_SCALAPACK
 #include "scalapack_vector_info.hpp"
 #endif
