+++ src/io_tests.cpp
@@ -1,7 +1,7 @@
 #include "io.hpp"
 
-#include "tensors.hpp"
 #include "tests_general.hpp"
+#include "vector.hpp"
 
 using namespace asgard;
