+++ src/moment.hpp
@@ -8,10 +8,11 @@ class PDE;
 
 #include "basis.hpp"
 #include "elements.hpp"
+#include "matrix.hpp"
 #include "pde/pde_base.hpp"
 #include "program_options.hpp"
 #include "sparse.hpp"
-#include "tensors.hpp"
+#include "vector.hpp"
 #include <vector>
 
 namespace asgard
