+++ src/pde/pde_continuity1.hpp
@@ -9,7 +9,6 @@
 #include <typeinfo>
 #include <vector>
 
-#include "../tensors.hpp"
 #include "pde_base.hpp"
 
 namespace asgard
