+++ src/pde/pde_fokkerplanck1_4p4.hpp
@@ -9,7 +9,6 @@
 #include <typeinfo>
 #include <vector>
 
-#include "../tensors.hpp"
 #include "pde_base.hpp"
 
 namespace asgard
