+++ src/sparse.hpp
@@ -7,8 +7,9 @@
 #endif
 
 #include "lib_dispatch.hpp"
-#include "tensors.hpp"
+#include "matrix.hpp"
 #include "tools.hpp"
+#include "vector.hpp"
 
 #include <filesystem>
 #include <memory>
