+++ src/tensors_tests.cpp
@@ -1,5 +1,6 @@
-#include "tensors.hpp"
+#include "matrix.hpp"
 #include "tests_general.hpp"
+#include "vector.hpp"
 #include <fstream>
 #include <functional>
 #include <iostream>
