+++ src/transformations.hpp
@@ -6,10 +6,11 @@
 #include "distribution.hpp"
 #include "elements.hpp"
 #include "fast_math.hpp"
+#include "matrix.hpp"
 #include "pde.hpp"
 #include "quadrature.hpp"
-#include "tensors.hpp"
 #include "tools.hpp"
+#include "vector.hpp"
 
 #include <algorithm>
 #include <cmath>
